+++ lizard_auth_client/migrations/0007_auto_20181116_1015.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('lizard_auth_client', '0006_auto_20181115_1313'),
+    ]
+
+    operations = [
+        migrations.AlterModelOptions(
+            name='organisation',
+            options={'ordering': ['name']},
+        ),
+    ]